Finalized Normal Math
I think this is done and working (I compared two ways of calculating tangents, one with the complex math and one with test and check. they looked the same, so I went with the complex way to make it look like we're doing more lol). If this looks good to other people, I think its ready to merge. Also adjusted the floor/ceiliing slightly, they were too high/low by a bit